requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.1
reportlab>=4.2.2
openpyxl>=3.1.5
qrcode[pil]>=7.4.2
Pillow>=10.3.0
